Feat: Create Select component
